fix: changed ViewUtils
